requirements.txt
alembic==1.4.3
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
confusable-homoglyphs==3.2.0
cryptography==3.3.1
defusedxml==0.7.0rc1
diff-match-patch==20200713
distlib==0.3.1
Django==3.1.3
django-allauth==0.44.0
django-appconf==1.0.4
django-bootstrap==0.2.4
django-bootstrap-v5==1.0.4
django-bootstrap3==15.0.0
django-bootstrap4==2.3.1
django-bootstrap5==1.0.0
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-docrootcms==2.0
django-environ==0.4.5
django-filter==2.4.0
django-google-charts==0.1.1
django-grappelli==2.14.3
django-import-export==2.5.0
django-js-asset==1.2.2
django-phonenumber-field==5.1.0
django-psycopg2-extension==0.0.5
django-qsstats-magic==1.1.0
django-ranged-response==0.2.0
django-registration==3.1.2
django-simple-captcha==0.5.13
django-user-accounts==3.0.4
djangorestframework==3.12.2
easy-thumbnails==2.7.1
et-xmlfile==1.1.0
filelock==3.0.12
idna==2.10
importlib-metadata==3.10.1
isort==5.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
pbr==5.5.0
phonenumbers==8.12.22
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.4.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.1
SQLAlchemy==1.3.19
sqlparse==0.4.1
stevedore==3.2.2
tablib==3.0.0
toml==0.10.1
urllib3==1.25.10
virtualenv==20.0.32
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1